[docs] Apply the new DX to the Autocomplete Field demo
