--- docs/data/date-pickers/custom-field/PickerWithJoyField.tsx.preview
@@ -1,12 +1,0 @@
-<MaterialCssVarsProvider>
-  <CssVarsProvider theme={{ [THEME_ID]: joyTheme }}>
-    <SyncThemeMode mode={materialTheme.palette.mode} />
-    <LocalizationProvider dateAdapter={AdapterDayjs}>
-      <JoyDatePicker
-        slotProps={{
-          field: { clearable: true },
-        }}
-      />
-    </LocalizationProvider>
-  </CssVarsProvider>
-</MaterialCssVarsProvider>
