+++ docs/data/date-pickers/custom-field/behavior-autocomplete/MaterialDatePicker.tsx
@@ -1,26 +1,49 @@
 import * as React from 'react';
-import dayjs from 'dayjs';
+import dayjs, { Dayjs } from 'dayjs';
 import Autocomplete from '@mui/material/Autocomplete';
 import TextField from '@mui/material/TextField';
 import Stack from '@mui/material/Stack';
 import { AdapterDayjs } from '@mui/x-date-pickers/AdapterDayjs';
 import { LocalizationProvider } from '@mui/x-date-pickers/LocalizationProvider';
-import { DatePicker } from '@mui/x-date-pickers/DatePicker';
+import {
+  DatePicker,
+  DatePickerFieldProps,
+  DatePickerProps,
+} from '@mui/x-date-pickers/DatePicker';
+import { useSplitFieldProps } from '@mui/x-date-pickers/hooks';
+import { useValidation, validateDate } from '@mui/x-date-pickers/validation';
 
-function AutocompleteField(props) {
+interface AutocompleteFieldProps extends DatePickerFieldProps<Dayjs> {
+  /**
+   * @typescript-to-proptypes-ignore
+   */
+  options?: Dayjs[];
+}
+
+function AutocompleteField(props: AutocompleteFieldProps) {
+  const { internalProps, forwardedProps } = useSplitFieldProps(props, 'date');
+  const { value, timezone, onChange } = internalProps;
   const {
+    InputProps,
+    slotProps,
+    slots,
+    ownerState,
     label,
-    disabled,
-    readOnly,
-    id,
-    value,
-    onChange,
-    InputProps: { ref, startAdornment, endAdornment } = {},
-    inputProps,
+    focused,
+    name,
     options = [],
-  } = props;
+    inputProps,
+    ...other
+  } = forwardedProps;
 
-  const mergeAdornments = (...adornments) => {
+  const { hasValidationError } = useValidation({
+    validator: validateDate,
+    value,
+    timezone,
+    props: internalProps,
+  });
+
+  const mergeAdornments = (...adornments: React.ReactNode[]) => {
     const nonNullAdornments = adornments.filter((el) => el != null);
     if (nonNullAdornments.length === 0) {
       return null;
@@ -41,25 +64,24 @@ function AutocompleteField(props) {
 
   return (
     <Autocomplete
-      id={id}
+      {...other}
       options={options}
-      disabled={disabled}
-      readOnly={readOnly}
-      ref={ref}
+      ref={InputProps?.ref}
       sx={{ minWidth: 250 }}
       renderInput={(params) => (
         <TextField
           {...params}
+          error={hasValidationError}
           label={label}
           inputProps={{ ...params.inputProps, ...inputProps }}
           InputProps={{
             ...params.InputProps,
             startAdornment: mergeAdornments(
-              startAdornment,
+              InputProps?.startAdornment,
               params.InputProps.startAdornment,
             ),
             endAdornment: mergeAdornments(
-              endAdornment,
+              InputProps?.endAdornment,
               params.InputProps.endAdornment,
             ),
           }}
@@ -83,22 +105,32 @@ function AutocompleteField(props) {
   );
 }
 
-function AutocompleteDatePicker(props) {
+interface AutocompleteDatePickerProps extends DatePickerProps<Dayjs> {
+  /**
+   * @typescript-to-proptypes-ignore
+   */
+  options: Dayjs[];
+}
+
+function AutocompleteDatePicker(props: AutocompleteDatePickerProps) {
   const { options, ...other } = props;
 
   const optionsLookup = React.useMemo(
     () =>
-      options.reduce((acc, option) => {
-        acc[option.toISOString()] = true;
-        return acc;
-      }, {}),
+      options.reduce(
+        (acc, option) => {
+          acc[option.toISOString()] = true;
+          return acc;
+        },
+        {} as Record<string, boolean>,
+      ),
     [options],
   );
 
   return (
     <DatePicker
       slots={{ ...props.slots, field: AutocompleteField }}
-      slotProps={{ ...props.slotProps, field: { options } }}
+      slotProps={{ ...props.slotProps, field: { options } as any }}
       shouldDisableDate={(date) => !optionsLookup[date.startOf('day').toISOString()]}
       {...other}
     />
@@ -107,7 +139,7 @@ function AutocompleteDatePicker(props) {
 
 const today = dayjs().startOf('day');
 
-export default function PickerWithAutocompleteField() {
+export default function MaterialDatePicker() {
   return (
     <LocalizationProvider dateAdapter={AdapterDayjs}>
       <AutocompleteDatePicker
